pyproject.toml
[project]
name = "my-rag-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.1",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
]
